+++ src/main/scala/scala/compat/java8/StepConverters.scala
@@ -8,102 +8,6 @@ import scala.compat.java8.runtime._
 package converterImpls {
   import Stepper._
 
-  trait MakesAnySeqStepper[A] extends Any {
-    def stepper: AnyStepper[A] 
-  }
-
-  trait MakesAnyKeySeqStepper[A] extends Any {
-    def keyStepper: AnyStepper[A] 
-  }
-
-  trait MakesAnyValueSeqStepper[A] extends Any {
-    def valueStepper: AnyStepper[A] 
-  }
-
-  trait MakesDoubleSeqStepper extends Any {
-    def stepper: DoubleStepper 
-  }
-
-  trait MakesDoubleKeySeqStepper extends Any {
-    def keyStepper: DoubleStepper 
-  }
-
-  trait MakesDoubleValueSeqStepper extends Any {
-    def valueStepper: DoubleStepper 
-  }
-
-  trait MakesIntSeqStepper extends Any {
-    def stepper: IntStepper 
-  }
-
-  trait MakesIntKeySeqStepper extends Any {
-    def keyStepper: IntStepper 
-  }
-
-  trait MakesIntValueSeqStepper extends Any {
-    def valueStepper: IntStepper 
-  }
-
-  trait MakesLongSeqStepper extends Any {
-    def stepper: LongStepper 
-  }
-
-  trait MakesLongKeySeqStepper extends Any {
-    def keyStepper: LongStepper 
-  }
-
-  trait MakesLongValueSeqStepper extends Any {
-    def valueStepper: LongStepper 
-  }
-
-  trait MakesAnyStepper[A] extends Any {
-    def stepper: AnyStepper[A] with EfficientSubstep
-  }
-
-  trait MakesAnyKeyStepper[A] extends Any {
-    def keyStepper: AnyStepper[A] with EfficientSubstep
-  }
-
-  trait MakesAnyValueStepper[A] extends Any {
-    def valueStepper: AnyStepper[A] with EfficientSubstep
-  }
-
-  trait MakesDoubleStepper extends Any {
-    def stepper: DoubleStepper with EfficientSubstep
-  }
-
-  trait MakesDoubleKeyStepper extends Any {
-    def keyStepper: DoubleStepper with EfficientSubstep
-  }
-
-  trait MakesDoubleValueStepper extends Any {
-    def valueStepper: DoubleStepper with EfficientSubstep
-  }
-
-  trait MakesIntStepper extends Any {
-    def stepper: IntStepper with EfficientSubstep
-  }
-
-  trait MakesIntKeyStepper extends Any {
-    def keyStepper: IntStepper with EfficientSubstep
-  }
-
-  trait MakesIntValueStepper extends Any {
-    def valueStepper: IntStepper with EfficientSubstep
-  }
-
-  trait MakesLongStepper extends Any {
-    def stepper: LongStepper with EfficientSubstep
-  }
-
-  trait MakesLongKeyStepper extends Any {
-    def keyStepper: LongStepper with EfficientSubstep
-  }
-
-  trait MakesLongValueStepper extends Any {
-    def valueStepper: LongStepper with EfficientSubstep
-  }
-    
   private[java8] class StepsObjectArray[A <: Object](underlying: Array[A], _i0: Int, _iN: Int)
   extends StepsLikeIndexed[A, StepsObjectArray[A]](_i0, _iN) {
     def next() = if (hasNext()) { val j = i0; i0 += 1; underlying(j) } else throwNSEE
